# digiwf-integrations/digiwf-oce-cosmos-integration/digiwf-oce-cosmos-integration-client/apidoc/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "The sysHUB REST API / SDK v3 is a RESTful API that provides third party teams access to sysHUB features through a standard way. This page documents all operations intended for third-party integration.",
    "version": "v3",
    "title": "sysHUB REST-API"
  },
  "host": "",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/webapi/v3/jobs/{id}/joblog": {
      "get": {
        "tags": [
          "Job log  operations"
        ],
        "summary": "Query job logs of one given job.",
        "description": "Get list of job logs featuring filtering (using RSQL), ordering and paging. Requires COSMOS permission 'PERM_IADMINSERVICE_GETLIST'.",
        "operationId": "getUsingGET_6",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the job owning retrieved job log entries",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Count of elements to retrieve from the given offset. MUST be used in conjunction with offset.",
            "required": false,
            "type": "integer",
            "minimum": 1.0,
            "exclusiveMinimum": false,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Index of the first element to retrieve. MUST be used in conjunction with limit.",
            "required": false,
            "type": "integer",
            "minimum": 0.0,
            "exclusiveMinimum": false,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "orderby",
            "in": "query",
            "description": "Sorting order of the queried result set having the following SQL-like syntax : attribute1[.asc|.desc][;attributeN[.asc|.desc]].",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "search",
            "in": "query",
            "description": "RSQL filtering search criteria.",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "The requested job log entries have been fetched and returned in the response body.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/JobLog"
              }
            }
          }
        },
        "security": [
          {
            "oauth2schema": [
              "public",
              "private"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "Job log  operations"
        ],
        "summary": "Create one job log for one given job.",
        "description": "Creates a new job log and receive the new job log data in the response body. Requires COSMOS permission 'PERM_IADMINSERVICE_ADDOBJECTID'.",
        "operationId": "postUsingPOST_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the job owning the job log to create",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "job log",
            "description": "representation of the job log to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JobLog"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "A new job log has been created out of the request body. The created job log entry has been sent in the response body.",
            "schema": {
              "$ref": "#/definitions/JobLog"
            }
          }
        },
        "security": [
          {
            "oauth2schema": [
              "public",
              "private"
            ]
          }
        ]
      }
    },
    "/webapi/v3/jobs/{id}/joblog/children": {
      "get": {
        "tags": [
          "Job log  operations"
        ],
        "summary": "Gets joblog children by given jobid and optional parentId for displaying joblog tree.",
        "description": "Gets joblog children by given jobid and optinal parentId for displaying joblog tree.",
        "operationId": "getChildrenUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the job",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "parentId",
            "in": "path",
            "description": "parentId of joblog",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "security": [
          {
            "oauth2schema": [
              "public",
              "private"
            ]
          }
        ]
      }
    },
    "/webapi/v3/jobs/{id}/joblog/children/{parentId}": {
      "get": {
        "tags": [
          "Job log  operations"
        ],
        "summary": "Gets joblog children by given jobid and optional parentId for displaying joblog tree.",
        "description": "Gets joblog children by given jobid and optinal parentId for displaying joblog tree.",
        "operationId": "getChildrenUsingGET_1",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the job",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "parentId",
            "in": "path",
            "description": "parentId of joblog",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "security": [
          {
            "oauth2schema": [
              "public",
              "private"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "oauth2schema": {
      "type": "oauth2",
      "tokenUrl": "",
      "flow": "password",
      "scopes": {
        "public": "public access",
        "private": "private access only for internal use"
      }
    }
  },
  "definitions": {
...
    "JobLog": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "example": 23,
          "description": "The unique ID of a JobLog object"
        },
        "jobId": {
          "type": "integer",
          "format": "int64",
          "example": 124,
          "description": "The unique ID of the job to which this JobLog object is assigned"
        },
        "logLevel": {
          "type": "integer",
          "format": "int64",
          "example": 10,
          "description": "The log level of this JobLog entry"
        },
        "message": {
          "type": "string",
          "example": "Successfully processed!",
          "description": "The message text"
        },
        "modifiedTime": {
          "type": "string",
          "format": "date-time",
          "example": "2018-07-13T16:20:38.707",
          "description": "Date/time that the job is to be deleted"
        },
        "parentId": {
          "type": "integer",
          "format": "int64",
          "example": 22,
          "description": "The ID of the parent JobLog entry"
        }
      },
      "title": "JobLog",
      "description": "Data object wrapping a COSMOS JobLog."
    },
...
  }
}